MAGE-1095: create BackendSearch Service to move getSearchResults method into
