+++ Helper/Data.php
@@ -9,6 +9,7 @@ use Algolia\AlgoliaSearch\Service\AdditionalSection\IndexBuilder as AdditionalSe
 use Algolia\AlgoliaSearch\Service\Category\IndexBuilder as CategoryIndexBuilder;
 use Algolia\AlgoliaSearch\Service\IndexNameFetcher;
 use Algolia\AlgoliaSearch\Service\Page\IndexBuilder as PageIndexBuilder;
+use Algolia\AlgoliaSearch\Service\Product\BackendSearch;
 use Algolia\AlgoliaSearch\Service\Product\IndexBuilder as ProductIndexBuilder;
 use Algolia\AlgoliaSearch\Service\Suggestion\IndexBuilder as SuggestionIndexBuilder;
 use Magento\Framework\Exception\LocalizedException;
@@ -28,6 +29,7 @@ class Data
         protected AdditionalSectionIndexBuilder $additionalSectionIndexBuilder,
         protected PageIndexBuilder              $pageIndexBuilder,
         protected SuggestionIndexBuilder        $suggestionIndexBuilder,
+        protected BackendSearch                 $backendSearch
     ){}
 
     /**
@@ -163,11 +165,11 @@ class Data
      * @internal This method is currently unstable and should not be used. It may be revisited or fixed in a future version.
      *
      * @deprecated
-     * Use Algolia\AlgoliaSearch\Service\Product\IndexBuilder::getSearchResult() instead
+     * Use Algolia\AlgoliaSearch\Service\Product\BackendSearch::getSearchResult() instead
      */
     public function getSearchResult(string $query, int $storeId, ?array $searchParams = null, ?string $targetedIndex = null): array
     {
-        return $this->productIndexBuilder->getSearchResult($query, $storeId, $searchParams, $targetedIndex);
+        return $this->backendSearch->getSearchResult($query, $storeId, $searchParams, $targetedIndex);
     }
 
     /**
